[wip] use diff3 like a module
