+++ Makefile
@@ -71,7 +71,6 @@ dist/iD.js: \
 	js/lib/d3.trigger.js \
 	js/lib/d3.curtain.js \
 	js/lib/d3.value.js \
-	js/lib/diff3.js \
 	js/lib/lodash.js \
 	js/id/start.js \
 	js/id/id.js \
