+++ test/index.html
@@ -34,7 +34,6 @@
     <script src='../js/lib/d3.value.js'></script>
     <script src='../js/lib/d3-compat.js'></script>
     <script src='../js/lib/bootstrap-tooltip.js'></script>
-    <script src='../js/lib/diff3.js'></script>
 
     <script src='../js/id/id.js'></script>
     <script src='../js/lib/id/index.js'></script>
