+++ test/index_packaged.html
@@ -30,7 +30,6 @@
 
     <!-- include spec files here... -->
     <script src='spec/lib/d3.keybinding.js'></script>
-    <script src='spec/lib/diff3.js'></script>
     <script src='spec/lib/locale.js'></script>
 
     <script src='spec/actions/add_entity.js'></script>
